adjusted subscription created email
