+++ packages/email/src/transactional.tsx
@@ -12,7 +12,7 @@ import NoEventsSecond from '../emails/sequences/no-events/second';
 import NoProjectFirst from '../emails/sequences/no-project/first';
 import NoProjectSecond from '../emails/sequences/no-project/second';
 import SubscriptionCancelledMail from '../emails/subscription-cancelled';
-import SubscriptionStartedMail from '../emails/subscription-started';
+import SubscriptionCreatedMail from '../emails/subscription-created';
 
 export const TRANSACTIONAL_FROM_EMAIL = 'Vemetric <info@vemetric.com>';
 export const TIPS_FROM_EMAIL = 'Vemetric <info@notifications.vemetric.com>';
@@ -60,9 +60,9 @@ export const TRANSACTIONAL_TEMPLATE_MAP = {
     subject: 'Thanks for being part of our journey!',
     email: SubscriptionCancelledMail,
   },
-  subscriptionStarted: {
+  subscriptionCreated: {
     subject: 'Welcome to Vemetric Pro!',
-    email: SubscriptionStartedMail,
+    email: SubscriptionCreatedMail,
   },
 };
 export type TemplateName = keyof typeof TRANSACTIONAL_TEMPLATE_MAP;
